move code to common base processor
